tests/justhtml-tests/character_preprocessing.dat

#data
<math><mi>\x00</mi></math>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,11): unexpected-null-character

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>

#data
<math><mi>\x00\x0c</mi></math>
#errors
(1,6): expected-doctype-but-got-start-tag
(1,11): unexpected-null-character

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         ""

#data
<math><mi>   </mi></math>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "   "

#data
<math><mi>\x0ctext</mi></math>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mi>
|         "text"

#data
<select>\x0ctext</select>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <select>
|       "text"

#data
<table>\x00</table>
#errors
(1,7): expected-doctype-but-got-start-tag
(1,8): unexpected-null-character

#document
| <html>
|   <head>
|   <body>
|     <table>

#data
<head></head>\x00
#errors
(1,6): expected-doctype-but-got-start-tag
(1,14): unexpected-null-character

#document
| <html>
|   <head>
|   <body>

#data
<table>\x0c</table>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <table>
|       ""

#data
<!doctype html>\x00
#errors
(1,16): unexpected-null-character

#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>


#data
<!--a\x00b-->
#errors
(1,1): expected-doctype-but-got-chars
(1,6): unexpected-null-character
#document
| <!-- a�b -->
| <html>
|   <head>
|   <body>

#data
<frameset>\x0d\x0a
#errors
#document
| <html>
|   <head>
|   <frameset>
|     "
"
#data
<!--a-
#errors
(1,1): expected-doctype-but-got-chars
(1,6): eof-in-comment
#document
| <!-- a -->
| <html>
|   <head>
|   <body>

#data
<!x\x0d\x0ay>
#errors
(1,1): expected-doctype-but-got-chars
#document
| <!-- x
y -->
| <html>
|   <head>
|   <body>

#data
<\x00div>
#errors
(1,1): expected-doctype-but-got-chars
(1,2): unexpected-null-character
#document
| <html>
|   <head>
|   <body>
|     "<�div>"

#data
</>
#errors
(1,1): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>

#data
\x00\x0cA
#errors
(1,1): unexpected-null-character
(1,1): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "A"

#data
<!DOCTYPE\x00html>
#errors
(1,10): unexpected-null-character
(1,15): unknown-doctype
#document
| <!DOCTYPE �html>
| <html>
|   <head>
|   <body>

#data
&#x;
#errors
(1,1): expected-doctype-but-got-chars
#document
| <html>
|   <head>
|   <body>
|     "&#x;"

#data
<!--first--><!DOCTYPE\x0d\x0ahtml>
#errors
(1,1): expected-doctype-but-got-chars
#document
| <!-- first -->
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<o\x0ce="\x0d">
#errors
(1,9): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <o>
|       e="
"

#data
<d 2="\x0d">
#errors
(1,9): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <d>
|       2="
"

#data
<a 8=&#13\x0d>
#errors
(1,11): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <a>
|       8="
"
